Recovered from 0145e55bcf6bd3e9b19f5c07819a1398b6a22c35

+++ src/datamodule.py
@@ -84,7 +84,7 @@ class ClayDataset(Dataset):
         cube = self.read_chip(chip_path)
 
         # remove nans and convert to tensor
-        cube["pixels"] = torch.as_tensor(data=cube["pixels"])
+        cube["pixels"] = torch.as_tensor(data=cube["pixels"], dtype=torch.float16)
         cube["bbox"] = torch.as_tensor(data=cube["bbox"], dtype=torch.float64)
         cube["epsg"] = torch.as_tensor(data=cube["epsg"], dtype=torch.int32)
         cube["date"] = str(cube["date"])
